forgot to rename dynamical DE header to specify it is tabulated
